Debugged main.py
There was an issue with the get_test_folder() function not being able to open the test_config file, even though the path was fine. As it turned out the issue was with the start_log() function that was called in the open file statement and had a corrupt path.