Work in progress on improved language detectioon
One feature of our app should be language detection, or at least generating the flashcards in the language specified by the user.
This hasn't been an issue so far, but the text currently used for testing sometimes contains English terminology and phrases. Still, it is mainly written in German and the flashcards, therefore, should also be in German. The model, however, tends to generates the flashcards in English.
I hoped that specifying the languages would resolve the issue. However, this showed little effect on the model output.
In the course of the next days I will experiment with the phrasing and positioning of the language instructions.