Successful flashcard generaion in other languages than English
Translating the whole prompt (including the system prompt and examples) worked. The flashcards are now consistently generated in German, with the same quality and quantity as in English.
Program restructuring as described in the last commit will follow soon.
Conclusion: LLMs, or GPT-3.5 at least, are bad at translating and performing complex tasks at the same time.
The implication: We need to ask the user for the language of the flashcards and/or use a seperete module for language detection.